+++ ios/Video/RCTVideoPlayerViewController.swift
@@ -1,13 +1,13 @@
 import AVKit
 
 class RCTVideoPlayerViewController: AVPlayerViewController {
-    
-    var rctDelegate:RCTVideoPlayerViewControllerDelegate!
-    
+
+    weak var rctDelegate: RCTVideoPlayerViewControllerDelegate?
+
     // Optional paramters
     var preferredOrientation:String?
     var autorotate:Bool?
-    
+
     func shouldAutorotate() -> Bool {
 
         if autorotate! || preferredOrientation == nil || (preferredOrientation!.lowercased() == "all") {
@@ -20,13 +20,11 @@ class RCTVideoPlayerViewController: AVPlayerViewController {
     override func viewDidDisappear(_ animated: Bool) {
         super.viewDidDisappear(animated)
 
-        if rctDelegate != nil {
-            rctDelegate.videoPlayerViewControllerWillDismiss(playerViewController: self)
-            rctDelegate.videoPlayerViewControllerDidDismiss(playerViewController: self)
-        }
+        rctDelegate?.videoPlayerViewControllerWillDismiss(playerViewController: self)
+        rctDelegate?.videoPlayerViewControllerDidDismiss(playerViewController: self)
     }
 
-    #if !TARGET_OS_TV
+    #if !os(tvOS)
 
     func supportedInterfaceOrientations() -> UIInterfaceOrientationMask {
         return .all
@@ -43,5 +41,6 @@ class RCTVideoPlayerViewController: AVPlayerViewController {
             return orientation
         }
     }
+    
     #endif
 }